initial whoop webhook implementation
