+++ .env.example
@@ -37,5 +37,3 @@ WHOOP_CLIENT_SECRET=your_whoop_client_secret
 WHOOP_API_URL=https://api.prod.whoop.com
 # Generate with: openssl rand -base64 32
 WHOOP_TOKEN_ENCRYPTION_KEY=your_32_byte_base64_encoded_key
-# Generate with: openssl rand -base64 32 (for webhook signature verification)
-WHOOP_WEBHOOK_SECRET=your_32_byte_base64_encoded_secret
