+++ drizzle/20260215021946_add_deleted_at_to_whoop_tables/migration.sql
@@ -1,0 +1,5 @@
+ALTER TABLE `whoop_recoveries` ADD `sleep_id` text;--> statement-breakpoint
+ALTER TABLE `whoop_recoveries` ADD `deleted_at` text;--> statement-breakpoint
+ALTER TABLE `whoop_sleeps` ADD `deleted_at` text;--> statement-breakpoint
+ALTER TABLE `whoop_webhook_events` ADD `trace_id` text;--> statement-breakpoint
+ALTER TABLE `whoop_workouts` ADD `deleted_at` text;
