+++ src/components/ui/EmptyState.tsx
@@ -1,18 +1,13 @@
-import { useState, useEffect, type ComponentType, type ReactNode } from 'react';
-import { Plus } from 'lucide-react';
 import { Button } from './Button';
-import { Card } from './Card';
+import type { ReactNode } from 'react';
 
 interface EmptyStateProps {
-  icon?: ComponentType<{ readonly size?: number; readonly className?: string }>;
+  icon?: (props: { readonly size?: number; readonly className?: string }) => ReactNode;
   title: string;
   description?: string;
   actionLabel?: string;
   onAction?: () => void;
-  secondaryActionLabel?: string;
-  onSecondaryAction?: () => void;
   className?: string;
-  variant?: 'card' | 'inline';
 }
 
 export function EmptyState({
@@ -21,35 +16,25 @@ export function EmptyState({
   description,
   actionLabel,
   onAction,
-  secondaryActionLabel,
-  onSecondaryAction,
   className = '',
-  variant = 'inline',
 }: EmptyStateProps): ReactNode {
-  const [mounted, setMounted] = useState(false);
   const IconComponent = icon;
 
-  useEffect(() => {
-    setMounted(true);
-  }, []);
-
-  const content = (
+  return (
     <div className={`text-center py-12 ${className}`}>
       <div className="flex items-center justify-center mb-6">
-        <div className={`flex h-16 w-16 items-center justify-center rounded-full bg-primary/10 ${mounted ? 'animate-in fade-in zoom-in duration-300' : ''}`}>
+        <div className="bg-primary/8 rounded-full p-5">
           {IconComponent ? (
-            <IconComponent size={56} className="text-primary/80" />
+            <IconComponent size={44} className="text-muted-foreground/50" />
           ) : (
             <svg
-              className="text-primary/80"
-              width={56}
-              height={56}
+              width={44}
+              height={44}
               viewBox="0 0 24 24"
               fill="none"
               stroke="currentColor"
               strokeWidth="1.5"
-              strokeLinecap="round"
-              strokeLinejoin="round"
+              className="text-muted-foreground/50"
             >
               <path d="M6.5 6.5a2.121 2.121 0 1 0 4.242 0 2.121 2.121 0 1 0-4.242 0Z" />
               <path d="M15.5 6.5a2.121 2.121 0 1 0 4.242 0 2.121 2.121 0 1 0-4.242 0Z" />
@@ -58,183 +43,129 @@ export function EmptyState({
           )}
         </div>
       </div>
-      <h3 className="text-lg font-semibold text-foreground mb-2 tracking-tight">{title}</h3>
+      <h3 className="text-lg font-semibold text-foreground mb-2">{title}</h3>
       {description ? (
-        <p className="text-muted-foreground mb-6 max-w-sm mx-auto text-sm leading-relaxed text-muted-foreground/80">
-          {description}
-        </p>
+        <p className="text-muted-foreground mb-6 max-w-sm mx-auto text-sm">{description}</p>
       ) : null}
-      {(actionLabel && onAction) || (secondaryActionLabel && onSecondaryAction) ? (
-        <div className="flex flex-col sm:flex-row items-center justify-center gap-3">
-          {actionLabel && onAction ? (
-            <Button onClick={onAction} size="lg">
-              <Plus size={18} />
-              {actionLabel}
-            </Button>
-          ) : null}
-          {secondaryActionLabel && onSecondaryAction ? (
-            <Button onClick={onSecondaryAction} variant="outline" size="lg">
-              {secondaryActionLabel}
-            </Button>
-          ) : null}
-        </div>
+      {actionLabel && onAction ? (
+        <Button onClick={onAction} size="lg" variant="cta">
+          {actionLabel}
+        </Button>
       ) : null}
     </div>
   );
-
-  if (variant === 'card') {
-    return <Card className="p-4">{content}</Card>;
-  }
-
-  return content;
 }
 
-export function EmptyTemplates({
-  searchActive,
-  onCreate,
-  onBrowse,
-}: {
-  readonly searchActive?: boolean;
-  readonly onCreate?: () => void;
-  readonly onBrowse?: () => void;
-}): ReactNode {
-  const FileTextIcon = ({ size = 56, className = '' }: { readonly size?: number; readonly className?: string }): ReactNode => (
-    <svg
-      className={className}
-      width={size}
-      height={size}
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M14.5 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7.5L14.5 2z" />
-      <polyline points="14 2 14 8 20 8" />
-      <line x1="16" x2="8" y1="13" y2="13" />
-      <line x1="16" x2="8" y1="17" y2="17" />
-      <line x1="10" x2="8" y1="9" y2="9" />
-    </svg>
-  );
+const DumbbellIcon = ({ size = 44, className = '' }: { readonly size?: number; readonly className?: string }): ReactNode => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="1.5"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <path d="M6.5 6.5a2.121 2.121 0 1 0 4.242 0 2.121 2.121 0 1 0-4.242 0Z" />
+    <path d="M15.5 6.5a2.121 2.121 0 1 0 4.242 0 2.121 2.121 0 1 0-4.242 0Z" />
+    <path d="M5 12h14" />
+  </svg>
+);
 
-  return (
-    <EmptyState
-      icon={FileTextIcon}
-      title="No Templates Yet"
-      description={
-        searchActive
-          ? "We couldn't find any templates matching your search. Try different keywords or clear your search."
-          : "Create your first workout template to organize your exercises and streamline your training sessions."
-      }
-      actionLabel={!searchActive ? 'Create Template' : undefined}
-      onAction={!searchActive ? onCreate : undefined}
-      secondaryActionLabel={searchActive ? 'Clear Search' : undefined}
-      onSecondaryAction={searchActive ? onBrowse : undefined}
-    />
-  );
-}
+const FileTextIcon = ({ size = 44, className = '' }: { readonly size?: number; readonly className?: string }): ReactNode => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="1.5"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <path d="M14.5 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7.5L14.5 2z" />
+    <polyline points="14 2 14 8 20 8" />
+    <line x1="16" x2="8" y1="13" y2="13" />
+    <line x1="16" x2="8" y1="17" y2="17" />
+    <line x1="10" x2="8" y1="9" y2="9" />
+  </svg>
+);
 
-export function EmptyExercises({
-  searchActive,
-  onCreate,
-  onBrowse,
-}: {
-  readonly searchActive?: boolean;
-  readonly onCreate?: () => void;
-  readonly onBrowse?: () => void;
-}): ReactNode {
-  const DumbbellIcon = ({ size = 56, className = '' }: { readonly size?: number; readonly className?: string }): ReactNode => (
-    <svg
-      className={className}
-      width={size}
-      height={size}
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M6.5 6.5a2.121 2.121 0 1 0 4.242 0 2.121 2.121 0 1 0-4.242 0Z" />
-      <path d="M15.5 6.5a2.121 2.121 0 1 0 4.242 0 2.121 2.121 0 1 0-4.242 0Z" />
-      <path d="M5 12h14" />
-    </svg>
-  );
+const ActivityIcon = ({ size = 44, className = '' }: { readonly size?: number; readonly className?: string }): ReactNode => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="1.5"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <polyline points="22 12 18 12 15 21 9 3 6 12 2 12" />
+  </svg>
+);
 
+const CalendarIcon = ({ size = 44, className = '' }: { readonly size?: number; readonly className?: string }): ReactNode => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="1.5"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <rect x="3" y="4" width="18" height="18" rx="2" ry="2" />
+    <line x1="16" x2="16" y1="2" y2="6" />
+    <line x1="8" x2="8" y1="2" y2="6" />
+    <line x1="3" y1="10" x2="21" y2="10" />
+  </svg>
+);
+
+export function EmptyExercises({ onCreate }: { readonly onCreate?: () => void }): ReactNode {
   return (
     <EmptyState
       icon={DumbbellIcon}
       title="No Exercises Yet"
-      description={
-        searchActive
-          ? "No exercises found matching your search. Try different keywords or browse all exercises."
-          : "Start building your exercise library by adding your first exercise. You can track strength, cardio, and flexibility exercises."
-      }
-      actionLabel={!searchActive ? 'Add Exercise' : undefined}
-      onAction={!searchActive ? onCreate : undefined}
-      secondaryActionLabel={searchActive ? 'View All' : undefined}
-      onSecondaryAction={searchActive ? onBrowse : undefined}
+      description="Start building your exercise library by adding your first exercise."
+      actionLabel="Add Exercise"
+      onAction={onCreate}
     />
   );
 }
 
-export function EmptyWorkouts({
-  onStart,
-  onBrowseTemplates,
-}: {
-  readonly onStart?: () => void;
-  readonly onBrowseTemplates?: () => void;
-}): ReactNode {
-  const ActivityIcon = ({ size = 56, className = '' }: { readonly size?: number; readonly className?: string }): ReactNode => (
-    <svg
-      className={className}
-      width={size}
-      height={size}
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <polyline points="22 12 18 12 15 21 9 3 6 12 2 12" />
-    </svg>
+export function EmptyTemplates({ onCreate }: { readonly onCreate?: () => void }): ReactNode {
+  return (
+    <EmptyState
+      icon={FileTextIcon}
+      title="No Templates Yet"
+      description="Create your first workout template to organize your exercises."
+      actionLabel="Create Template"
+      onAction={onCreate}
+    />
   );
+}
 
+export function EmptyWorkouts({ onStart }: { readonly onStart?: () => void }): ReactNode {
   return (
     <EmptyState
       icon={ActivityIcon}
       title="Ready to Train?"
-      description="You haven't completed any workouts yet. Start your fitness journey by beginning your first workout or choosing a template."
+      description="You haven't completed any workouts yet. Start your fitness journey by beginning your first workout."
       actionLabel="Start Workout"
       onAction={onStart}
-      secondaryActionLabel="Browse Templates"
-      onSecondaryAction={onBrowseTemplates}
     />
   );
 }
 
 export function EmptyExerciseHistory(): ReactNode {
-  const CalendarIcon = ({ size = 56, className = '' }: { readonly size?: number; readonly className?: string }): ReactNode => (
-    <svg
-      className={className}
-      width={size}
-      height={size}
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <rect x="3" y="4" width="18" height="18" rx="2" ry="2" />
-      <line x1="16" x2="16" y1="2" y2="6" />
-      <line x1="8" x2="8" y1="2" y2="6" />
-      <line x1="3" y1="10" x2="21" y2="10" />
-    </svg>
-  );
-
   return (
     <EmptyState
       icon={CalendarIcon}
@@ -243,5 +174,3 @@ export function EmptyExerciseHistory(): ReactNode {
     />
   );
 }
-
-
