+++ src/routes/api/exercises.ts
@@ -1,8 +1,8 @@
 import { createFileRoute } from '@tanstack/react-router';
 import { createExercise, getExercisesByWorkosId } from '../../lib/db/exercise';
 import { validateBody, MAX_SEARCH_TERM_LENGTH } from '~/lib/api/route-helpers';
-import { withApiContext } from '~/lib/api/context';
-import { createApiError, ApiError } from '~/lib/api/errors';
+import { withApiContext, parseQueryParams, handleApiError } from '~/lib/api/handler';
+import { createApiError, API_ERROR_CODES } from '~/lib/api/errors';
 import { createExerciseSchema } from '~/lib/validators';
 
 export const Route = createFileRoute('/api/exercises')({
@@ -13,13 +13,15 @@ export const Route = createFileRoute('/api/exercises')({
           const { session, d1Db } = await withApiContext(request);
 
           const url = new URL(request.url);
-          const search = url.searchParams.get('search') ?? undefined;
-          const muscleGroup = url.searchParams.get('muscleGroup') ?? undefined;
-          const sortBy = url.searchParams.get('sortBy') as 'createdAt' | 'muscleGroup' | 'name' | undefined;
-          const sortOrder = url.searchParams.get('sortOrder') as 'ASC' | 'DESC' | undefined;
+          const { search, muscleGroup, sortBy, sortOrder } = parseQueryParams<{
+            search?: string;
+            muscleGroup?: string;
+            sortBy?: 'createdAt' | 'muscleGroup' | 'name';
+            sortOrder?: 'ASC' | 'DESC';
+          }>(url);
 
           if (search && search.length > MAX_SEARCH_TERM_LENGTH) {
-            return createApiError('Search term too long', 400, 'VALIDATION_ERROR');
+            return createApiError('Search term too long', 400, API_ERROR_CODES.VALIDATION_ERROR);
           }
 
           const exercises = await getExercisesByWorkosId(d1Db, session.sub, {
@@ -31,11 +33,7 @@ export const Route = createFileRoute('/api/exercises')({
 
           return Response.json(exercises);
         } catch (err) {
-          if (err instanceof ApiError) {
-            return createApiError(err.message, err.status, err.code);
-          }
-          console.error('Get exercises error:', err);
-          return createApiError('Server error', 500, 'SERVER_ERROR');
+          return handleApiError(err, 'Get exercises');
         }
       },
       POST: async ({ request }) => {
@@ -44,7 +42,7 @@ export const Route = createFileRoute('/api/exercises')({
 
           const body = await validateBody(request, createExerciseSchema);
           if (!body) {
-            return createApiError('Invalid request body', 400, 'VALIDATION_ERROR');
+            return createApiError('Invalid request body', 400, API_ERROR_CODES.VALIDATION_ERROR);
           }
 
           const { name, muscleGroup, description, localId, libraryId } = body;
@@ -60,11 +58,7 @@ export const Route = createFileRoute('/api/exercises')({
 
           return Response.json(exercise, { status: 201 });
         } catch (err) {
-          if (err instanceof ApiError) {
-            return createApiError(err.message, err.status, err.code);
-          }
-          console.error('Create exercise error:', err);
-          return createApiError('Server error', 500, 'SERVER_ERROR');
+          return handleApiError(err, 'Create exercise');
         }
       },
     },
