+++ src/routes/api/user.1rm.ts
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router';
 import { withApiContext } from '../../lib/api/context';
-import { createApiError, ApiError } from '../../lib/api/errors';
+import { createApiError, ApiError, API_ERROR_CODES } from '../../lib/api/errors';
 import { getLatestOneRMs } from '~/lib/db/program';
 
 export const Route = createFileRoute('/api/user/1rm')({
@@ -27,7 +27,7 @@ export const Route = createFileRoute('/api/user/1rm')({
              return createApiError(err.message, err.status, err.code);
            }
            console.error('Get latest 1RM error:', err);
-           return createApiError('Server error', 500, 'SERVER_ERROR');
+           return createApiError('Server error', 500, API_ERROR_CODES.SERVER_ERROR);
          }
       },
     },
