+++ src/routes/api/webhooks.whoop/index.tsx
@@ -1,92 +1,109 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { env } from 'cloudflare:workers';
+import { env, waitUntil } from 'cloudflare:workers';
+import { timingSafeEqual } from 'node:crypto';
+import { Buffer } from 'node:buffer';
 import { whoopRepository, WhoopApiClient } from '~/lib/whoop';
 import { trackEvent } from '~/lib/posthog';
-import { mapWhoopSleepToDb, mapWhoopRecoveryToDb, mapWhoopCycleToDb, mapWhoopWorkoutToDb } from '~/lib/whoop/api';
+import { mapWhoopSleepToDb, mapWhoopRecoveryToDb, mapWhoopWorkoutToDb } from '~/lib/whoop/api';
 
-const WHOOP_WEBHOOK_SECRET = process.env.WHOOP_WEBHOOK_SECRET;
+const REPLAY_TOLERANCE_MS = 5 * 60 * 1000;
 
-function addDays(date: Date, days: number): Date {
-  const result = new Date(date);
-  result.setDate(result.getDate() + days);
-  return result;
-}
+async function verifyWebhookSignature(
+  timestamp: string,
+  body: string,
+  signature: string,
+  clientSecret: string
+): Promise<boolean> {
+  if (!clientSecret) {
+    return false;
+  }
 
-function formatDate(date: Date): string {
-  return date.toISOString();
-}
+  const timestampNum = parseInt(timestamp, 10);
+  if (isNaN(timestampNum)) {
+    return false;
+  }
 
-async function verifyWebhookSignature(body: string, signature: string): Promise<boolean> {
-  if (!WHOOP_WEBHOOK_SECRET) {
+  const now = Date.now();
+  if (Math.abs(now - timestampNum) > REPLAY_TOLERANCE_MS) {
     return false;
   }
 
   const encoder = new TextEncoder();
+  const data = timestamp + body;
+  const keyData = encoder.encode(clientSecret);
+  const messageData = encoder.encode(data);
+
   const key = await crypto.subtle.importKey(
     'raw',
-    encoder.encode(WHOOP_WEBHOOK_SECRET),
+    keyData,
     { name: 'HMAC', hash: 'SHA-256' },
     false,
-    ['verify']
+    ['sign']
   );
 
-  const signatureBuffer = Buffer.from(signature, 'hex');
-  const bodyBuffer = encoder.encode(body);
+  const signatureBytes = await crypto.subtle.sign('HMAC', key, messageData);
+  const expectedBuffer = Buffer.from(new Uint8Array(signatureBytes));
+  const sigBuffer = Buffer.from(signature, 'base64');
 
-  return crypto.subtle.verify('HMAC', key, signatureBuffer, bodyBuffer);
+  if (sigBuffer.length !== expectedBuffer.length) {
+    return false;
+  }
+
+  return timingSafeEqual(sigBuffer, expectedBuffer);
 }
 
-async function processWebhookEvent(d1Db: D1Database, payload: { event_id: string; event_type: string; user_id: number; payload: unknown }) {
+async function processWebhookEvent(
+  d1Db: D1Database,
+  payload: { user_id: number; id: string; type: string; trace_id: string }
+) {
+  const { user_id: userId, id, type, trace_id: traceId } = payload;
+
   try {
-    const connection = await whoopRepository.getConnectionByWhoopUserId(d1Db, payload.user_id.toString());
+    const connection = await whoopRepository.getConnectionByWhoopUserId(d1Db, userId.toString());
     if (!connection) {
-      await whoopRepository.markWebhookProcessed(d1Db, payload.event_id, 'No connection found for whoop user');
+      await whoopRepository.markWebhookProcessed(d1Db, traceId, 'No connection found for whoop user');
       return;
     }
 
     const workosId = connection.workosId;
     const client = new WhoopApiClient(workosId, d1Db);
-    const now = new Date();
-    const endDate = formatDate(now);
-    const startDate = formatDate(addDays(now, -7));
 
-    switch (payload.event_type) {
-      case 'daily.data.updated': {
-        const cycles = await client.getCycles(startDate, endDate);
-        for (const cycle of cycles.records) {
-          await whoopRepository.upsertCycle(d1Db, workosId, mapWhoopCycleToDb(cycle, workosId));
-        }
-        const recoveries = await client.getRecoveries(startDate, endDate);
-        for (const recovery of recoveries.records) {
-          await whoopRepository.upsertRecovery(d1Db, workosId, mapWhoopRecoveryToDb(recovery, workosId));
-        }
-        const sleeps = await client.getSleeps(startDate, endDate);
-        for (const sleep of sleeps.records) {
-          await whoopRepository.upsertSleep(d1Db, workosId, mapWhoopSleepToDb(sleep, workosId));
-        }
+    switch (type) {
+      case 'workout.updated': {
+        const workout = await client.getWorkoutById(id);
+        await whoopRepository.upsertWorkout(d1Db, workosId, mapWhoopWorkoutToDb(workout, workosId));
         break;
       }
-      case 'workout.completed': {
-        const workouts = await client.getWorkouts(startDate, endDate);
-        for (const workout of workouts.records) {
-          await whoopRepository.upsertWorkout(d1Db, workosId, mapWhoopWorkoutToDb(workout, workosId));
-        }
+      case 'workout.deleted': {
+        await whoopRepository.updateWorkoutDeletedAt(d1Db, workosId, id, new Date().toISOString());
         break;
       }
-      case 'sleep.completed': {
-        const sleeps = await client.getSleeps(startDate, endDate);
-        for (const sleep of sleeps.records) {
-          await whoopRepository.upsertSleep(d1Db, workosId, mapWhoopSleepToDb(sleep, workosId));
-        }
+      case 'sleep.updated': {
+        const sleep = await client.getSleepById(id);
+        await whoopRepository.upsertSleep(d1Db, workosId, mapWhoopSleepToDb(sleep, workosId));
+        break;
+      }
+      case 'sleep.deleted': {
+        await whoopRepository.updateSleepDeletedAt(d1Db, workosId, id, new Date().toISOString());
+        await whoopRepository.updateRecoveryDeletedAt(d1Db, workosId, id, new Date().toISOString());
+        break;
+      }
+      case 'recovery.updated': {
+        const recovery = await client.getRecoveryBySleepId(id);
+        await whoopRepository.upsertRecovery(d1Db, workosId, mapWhoopRecoveryToDb(recovery, workosId));
+        break;
+      }
+      case 'recovery.deleted': {
+        await whoopRepository.updateRecoveryDeletedAt(d1Db, workosId, id, new Date().toISOString());
         break;
       }
     }
 
     await whoopRepository.updateLastSyncAt(d1Db, workosId);
-    await whoopRepository.markWebhookProcessed(d1Db, payload.event_id);
+    await whoopRepository.markWebhookProcessed(d1Db, traceId);
   } catch (error) {
     const errorMessage = error instanceof Error ? error.message : 'Unknown error';
-    await whoopRepository.markWebhookProcessed(d1Db, payload.event_id, errorMessage);
+    await whoopRepository.markWebhookProcessed(d1Db, traceId, errorMessage);
   }
 }
 
@@ -94,19 +111,34 @@ export const Route = createFileRoute('/api/webhooks/whoop/')({
   server: {
     handlers: {
       POST: async ({ request }) => {
+        const timestamp = request.headers.get('X-WHOOP-Signature-Timestamp') ?? '';
+        const signature = request.headers.get('X-WHOOP-Signature') ?? '';
         const body = await request.text();
-        const signature = request.headers.get('X-Whoop-Signature') ?? '';
 
-        const isValid = await verifyWebhookSignature(body, signature);
+        const clientSecret = (env as { WHOOP_CLIENT_SECRET?: string }).WHOOP_CLIENT_SECRET ?? '';
+        const isValid = await verifyWebhookSignature(timestamp, body, signature, clientSecret);
         if (!isValid) {
           return new Response('Invalid signature', { status: 401 });
         }
 
-        let payload: { event_id: string; event_type: string; user_id: number; payload: unknown };
+        let payload: { user_id: number; id: string; type: string; trace_id: string };
         try {
-          payload = JSON.parse(body) as { event_id: string; event_type: string; user_id: number; payload: unknown };
+          payload = JSON.parse(body) as { user_id: number; id: string; type: string; trace_id: string };
         } catch {
           return new Response('Invalid JSON', { status: 400 });
+        }
+
+        const validEventTypes = [
+          'workout.updated',
+          'workout.deleted',
+          'sleep.updated',
+          'sleep.deleted',
+          'recovery.updated',
+          'recovery.deleted',
+        ];
+
+        if (!validEventTypes.includes(payload.type)) {
+          return Response.json({ status: 'ignored' });
         }
 
         const d1Db = (env as { DB?: D1Database }).DB;
@@ -115,8 +147,9 @@ export const Route = createFileRoute('/api/webhooks/whoop/')({
         }
 
         const webhookEvent = {
-          id: payload.event_id,
-          eventType: payload.event_type,
+          id: payload.trace_id,
+          traceId: payload.trace_id,
+          eventType: payload.type,
           payloadRaw: body,
         };
 
@@ -125,9 +158,9 @@ export const Route = createFileRoute('/api/webhooks/whoop/')({
           return Response.json({ status: 'duplicate' });
         }
 
-        await trackEvent('whoop_webhook_received', { eventType: payload.event_type });
+        await trackEvent('whoop_webhook_received', { eventType: payload.type });
 
-        void processWebhookEvent(d1Db, payload);
+        waitUntil(processWebhookEvent(d1Db, payload));
 
         return Response.json({ status: 'received' });
       },
