+++ src/routes/api/workouts.$id.exercises.ts
@@ -7,7 +7,7 @@ import {
   reorderWorkoutExercises
 } from '../../lib/db/workout';
 import { withApiContext } from '../../lib/api/context';
-import { createApiError } from '../../lib/api/errors';
+import { createApiError, API_ERROR_CODES } from '../../lib/api/errors';
 
 export const Route = createFileRoute('/api/workouts/$id/exercises')({
   server: {
@@ -20,7 +20,7 @@ export const Route = createFileRoute('/api/workouts/$id/exercises')({
             return Response.json(exercises);
           } catch (err) {
             console.error('Get workout exercises error:', err);
-            return createApiError('Server error', 500, 'SERVER_ERROR');
+            return createApiError('Server error', 500, API_ERROR_CODES.SERVER_ERROR);
           }
         },
        POST: async ({ request, params }) => {
@@ -31,7 +31,7 @@ export const Route = createFileRoute('/api/workouts/$id/exercises')({
             const { exerciseId, orderIndex, notes, localId } = body as { exerciseId: string; orderIndex: number; notes?: string; localId?: string };
 
             if (!exerciseId || orderIndex === undefined) {
-              return createApiError('Exercise ID and order index are required', 400, 'VALIDATION_ERROR');
+              return createApiError('Exercise ID and order index are required', 400, API_ERROR_CODES.VALIDATION_ERROR);
             }
 
             const workoutExercise = await createWorkoutExercise(
@@ -45,13 +45,13 @@ export const Route = createFileRoute('/api/workouts/$id/exercises')({
             );
 
             if (!workoutExercise) {
-              return createApiError('Workout not found or does not belong to you', 404, 'NOT_FOUND');
+              return createApiError('Workout not found or does not belong to you', 404, API_ERROR_CODES.NOT_FOUND);
             }
 
             return Response.json(workoutExercise, { status: 201 });
           } catch (err) {
             console.error('Add exercise error:', err);
-            return createApiError('Server error', 500, 'SERVER_ERROR');
+            return createApiError('Server error', 500, API_ERROR_CODES.SERVER_ERROR);
           }
         },
        DELETE: async ({ request, params }) => {
@@ -62,19 +62,19 @@ export const Route = createFileRoute('/api/workouts/$id/exercises')({
             const exerciseId = url.searchParams.get('exerciseId');
 
             if (!exerciseId) {
-              return createApiError('Exercise ID is required', 400, 'VALIDATION_ERROR');
+              return createApiError('Exercise ID is required', 400, API_ERROR_CODES.VALIDATION_ERROR);
             }
 
             const removed = await removeWorkoutExercise(d1Db, params.id, exerciseId, session.sub);
 
             if (!removed) {
-              return createApiError('Exercise not found', 404, 'NOT_FOUND');
+              return createApiError('Exercise not found', 404, API_ERROR_CODES.NOT_FOUND);
             }
 
             return new Response(null, { status: 204 });
           } catch (err) {
             console.error('Remove exercise error:', err);
-            return createApiError('Server error', 500, 'SERVER_ERROR');
+            return createApiError('Server error', 500, API_ERROR_CODES.SERVER_ERROR);
           }
       },
       PUT: async ({ request, params }) => {
@@ -85,19 +85,19 @@ export const Route = createFileRoute('/api/workouts/$id/exercises')({
           const { exerciseOrders } = body as { exerciseOrders: ExerciseOrder[] };
 
           if (!exerciseOrders || !Array.isArray(exerciseOrders)) {
-            return createApiError('Exercise orders array is required', 400, 'VALIDATION_ERROR');
+            return createApiError('Exercise orders array is required', 400, API_ERROR_CODES.VALIDATION_ERROR);
           }
 
           const reordered = await reorderWorkoutExercises(db, params.id, exerciseOrders, session.sub);
 
           if (!reordered) {
-            return createApiError('Workout not found', 404, 'NOT_FOUND');
+            return createApiError('Workout not found', 404, API_ERROR_CODES.NOT_FOUND);
           }
 
           return Response.json({ success: true });
         } catch (err) {
           console.error('Reorder exercises error:', err);
-            return createApiError('Server error', 500, 'SERVER_ERROR');
+            return createApiError('Server error', 500, API_ERROR_CODES.SERVER_ERROR);
           }
         },
