+++ src/routes/api/workouts.sets.$setId.ts
@@ -1,7 +1,7 @@
 import { createFileRoute } from '@tanstack/react-router';
 import { type NewWorkoutSet, deleteWorkoutSet, updateWorkoutSet } from '../../lib/db/workout';
 import { withApiContext } from '../../lib/api/context';
-import { createApiError } from '../../lib/api/errors';
+import { createApiError, API_ERROR_CODES } from '../../lib/api/errors';
 
 export const Route = createFileRoute('/api/workouts/sets/$setId')({
   server: {
@@ -12,7 +12,7 @@ export const Route = createFileRoute('/api/workouts/sets/$setId')({
 
             if (!params.setId || typeof params.setId !== 'string') {
               console.error('Invalid set ID:', params.setId);
-              return createApiError('Invalid set ID', 400, 'VALIDATION_ERROR');
+              return createApiError('Invalid set ID', 400, API_ERROR_CODES.VALIDATION_ERROR);
             }
 
             const body = await request.json();
@@ -37,28 +37,28 @@ export const Route = createFileRoute('/api/workouts/sets/$setId')({
 
             if (weight !== undefined && weight !== null) {
               if (typeof weight !== 'number' || weight < 0) {
-                return createApiError('Weight must be a non-negative number', 400, 'VALIDATION_ERROR');
+                return createApiError('Weight must be a non-negative number', 400, API_ERROR_CODES.VALIDATION_ERROR);
               }
               updateData.weight = weight;
             }
 
             if (reps !== undefined && reps !== null) {
               if (typeof reps !== 'number' || reps < 0) {
-                return createApiError('Reps must be a non-negative number', 400, 'VALIDATION_ERROR');
+                return createApiError('Reps must be a non-negative number', 400, API_ERROR_CODES.VALIDATION_ERROR);
               }
               updateData.reps = reps;
             }
 
             if (rpe !== undefined && rpe !== null) {
               if (typeof rpe !== 'number' || rpe < 0) {
-                return createApiError('RPE must be a non-negative number', 400, 'VALIDATION_ERROR');
+                return createApiError('RPE must be a non-negative number', 400, API_ERROR_CODES.VALIDATION_ERROR);
               }
               updateData.rpe = rpe;
             }
 
             if (isComplete !== undefined) {
               if (typeof isComplete !== 'boolean') {
-                return createApiError('isComplete must be a boolean', 400, 'VALIDATION_ERROR');
+                return createApiError('isComplete must be a boolean', 400, API_ERROR_CODES.VALIDATION_ERROR);
               }
               updateData.isComplete = isComplete;
               if (isComplete) {
@@ -67,20 +67,20 @@ export const Route = createFileRoute('/api/workouts/sets/$setId')({
             }
 
             if (Object.keys(updateData).length === 0) {
-              return createApiError('No valid fields to update', 400, 'VALIDATION_ERROR');
+              return createApiError('No valid fields to update', 400, API_ERROR_CODES.VALIDATION_ERROR);
             }
 
             const workoutSet = await updateWorkoutSet(d1Db, params.setId, session.sub, updateData as Partial<NewWorkoutSet>);
 
             if (!workoutSet) {
                console.warn('Set not found or does not belong to user:', { setId: params.setId, workosId: session.sub });
-              return createApiError('Set not found or does not belong to you', 404, 'NOT_FOUND');
+              return createApiError('Set not found or does not belong to you', 404, API_ERROR_CODES.NOT_FOUND);
             }
 
              return Response.json(workoutSet);
           } catch (err) {
             console.error('Update set error:', err);
-            return createApiError('Server error', 500, 'SERVER_ERROR');
+            return createApiError('Server error', 500, API_ERROR_CODES.SERVER_ERROR);
           }
         },
        DELETE: async ({ request, params }) => {
@@ -89,20 +89,20 @@ export const Route = createFileRoute('/api/workouts/sets/$setId')({
 
             if (!params.setId || typeof params.setId !== 'string') {
               console.error('Invalid set ID for delete:', params.setId);
-              return createApiError('Invalid set ID', 400, 'VALIDATION_ERROR');
+              return createApiError('Invalid set ID', 400, API_ERROR_CODES.VALIDATION_ERROR);
             }
 
             const deleted = await deleteWorkoutSet(d1Db, params.setId, session.sub);
 
             if (!deleted) {
               console.warn('Delete set failed - not found or does not belong to user:', { setId: params.setId, workosId: session.sub });
-              return createApiError('Set not found or does not belong to you', 404, 'NOT_FOUND');
+              return createApiError('Set not found or does not belong to you', 404, API_ERROR_CODES.NOT_FOUND);
             }
 
             return new Response(null, { status: 204 });
           } catch (err) {
             console.error('Delete set error:', err);
-            return createApiError('Server error', 500, 'SERVER_ERROR');
+            return createApiError('Server error', 500, API_ERROR_CODES.SERVER_ERROR);
           }
         },
     },
